bcs-app/requirements.txt
Django==1.11.5
djangorestframework==3.6.4
pytz==2017.2
# Use lower version of PyMySQL to connect to GCS MySQL successfully
PyMySQL==0.6.7

bleach==2.0.0
enum34==1.1.6

redis==2.10.5
python-logstash==0.4.6
dogpile.cache==0.6.4

six==1.11.0
requests==2.18.4
arrow==0.10.0
django-cors-headers==2.1.0

mako==1.0.7
jsonschema==2.6.0
jsonfield==2.0.2

gevent==1.2.2
django-redis==4.8.0
whitenoise==3.3.1

gunicorn==19.7.1
PyJWT==1.5.3
cryptography==2.1.4
python-dateutil==2.6.1
PyYAML==3.12
django-extensions==2.0.7
werkzeug==0.14.1
Pillow==5.1.0
dataclasses==0.6
django-yamlfield==1.0.3
Jinja2==2.10
boto==2.49.0
botocore==1.10.84
furl==2.0.0
dpath==1.4.2
tornado==5.1.1
kubernetes==7.0.0
ruamel.yaml==0.15.79
ruamel.yaml.cmd==0.5.3

pycrypto==2.6.1
django-picklefield==1.1.0

celery==3.1.18
django-celery==3.2.1

natsort==5.4.1
raven==6.0.0